move negotation helpers to negotiate.go
